Min-cut

GraphsFlows.mincut — Method

mincut(flow_graph::Graphs.IsDirected, source::Integer, target::Integer, capacity_matrix::AbstractMatrix, algorithm::AbstractFlowAlgorithm)

Compute the min-cut between source and target for the given graph. First computes the maxflow using algorithm and then builds the partition of the residual graph Returns a triplet (part1, part2, flow) with the partition containing the source, the partition containing the target (the rest) and the min-cut(max-flow) value